renaming and repositioning files
