+++ lib/app/app.dart
@@ -1,0 +1,2 @@
+export 'view/widgets/widgets.dart';
+export 'counter_app.dart';
