+++ lib/app/view/widgets/button.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+part of 'widgets.dart';
 
 class Btn extends StatelessWidget {
   const Btn(
