+++ lib/app/view/widgets/widgets.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+part 'button.dart';
+part 'text_form_field.dart';
