+++ lib/counter/controller/controller.dart
@@ -1,1 +1,1 @@
-part 'counterController.dart';
+part 'counter_controller.dart';
