+++ lib/counter/view/view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:counter2point0/counter/counter.dart';
 import 'package:counter2point0/theme/theme.dart';
-import 'package:counter2point0/widgets/widgets.dart';
+import 'package:counter2point0/app/app.dart';
 
-part 'editableCounter.dart';
-part 'buttonsRow.dart';
-part 'counterMain.dart';
+part 'editable_counter.dart';
+part 'buttons_row.dart';
+part 'counter_main.dart';
