+++ lib/main.dart
@@ -1,8 +1,8 @@
-import 'package:counter2point0/app/root.dart';
+import 'package:counter2point0/app/counter_app.dart';
 import 'package:flutter/material.dart';
 
 void main() {
   runApp(
-    const MyApp(),
+    const CounterApp(),
   );
 }
