+++ lib/theme/controller/controller.dart
@@ -1,3 +1,3 @@
 import 'package:flutter/material.dart';
 
-part 'themeController.dart';
+part 'theme_controller.dart';
