+++ lib/theme/view/view.dart
@@ -1,4 +1,4 @@
 import 'package:counter2point0/theme/theme.dart';
 import 'package:flutter/material.dart';
 
-part 'themeButton.dart';
+part 'theme_button.dart';
